refactor(bootstrap): split lib.rs into config/built/builder/download_trigger modules
Phase 0c of #458. Pure restructure — no behavior change.

The single ~290-line lib.rs is split into:
- config.rs       — BootstrapConfig
- built.rs        — BuiltCore
- builder.rs      — CoreBootstrap + build()
- download_trigger.rs — pub(crate) DownloadTriggerAdapter
- lib.rs          — crate-level docs and re-exports

Each module stays small and focused. Public surface unchanged
(BootstrapConfig, BuiltCore, CoreBootstrap re-exported from lib.rs).
